#1: update take book flow to support correct response fmt

+++ v4/src/apiserver/core/ports/rating/models.go
@@ -1,5 +1,5 @@
 package rating
 
 type Rating struct {
-	Stars uint32
+	Stars uint64
 }
